# openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The purpose of this service is to properly track daily activities.",
        "title": "dailytrack",
        "contact": {
            "name": "DailyTrack",
            "email": "help@512b.it"
        },
        "version": "0.0.1"
    },
    "host": "localhost:3000",
    "paths": {
        "/v1/api_keys": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "List all api keys for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api_key"
                ],
                "summary": "List api keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.List-models_ApiKey"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Create an api key for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api_key"
                ],
                "summary": "Create an api key",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "Name of the api key",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ApiKey"
                        }
                    }
                }
            }
        },
        "/v1/events/{username}/{track_name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all events for a track",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "List events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Track name",
                        "name": "track_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Created at",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantity",
                        "name": "quantity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.List-models_Day"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create an event for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Create an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Track name",
                        "name": "track_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Created at",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantity",
                        "name": "quantity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModel"
                        }
                    }
                }
            }
        },
        "/v1/tracks": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "List all tracks for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "track"
                ],
                "summary": "List tracks",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.List-models_Track"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Create a track",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "track"
                ],
                "summary": "Create a track",
                "parameters": [
                    {
                        "description": "Track object",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Track"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModel"
                        }
                    }
                }
            }
        },
        "/v1/users/challenge": {
            "get": {
                "description": "Create a user challenge given an email address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create a user challenge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The email address of the user",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Challenge"
                        }
                    }
                }
            }
        },
        "/v1/users/signin": {
            "post": {
                "description": "Sign in a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Sign in a user",
                "parameters": [
                    {
                        "description": "the signed challenge to enter the system",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignIn"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModel"
                        }
                    }
                }
            }
        },
        "/v1/users/signup": {
            "post": {
                "description": "Create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "the user body to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ApiKey": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "delete_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Challenge": {
            "type": "object",
            "properties": {
                "challenge": {
                    "type": "string"
                },
                "salt": {
                    "type": "string"
                }
            }
        },
        "models.Day": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "models.List-models_ApiKey": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ApiKey"
                    }
                }
            }
        },
        "models.List-models_Day": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Day"
                    }
                }
            }
        },
        "models.List-models_Track": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Track"
                    }
                }
            }
        },
        "models.ResponseModel": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "email is invalid"
                },
                "message": {
                    "type": "string",
                    "example": "invalid"
                }
            }
        },
        "models.SignIn": {
            "type": "object",
            "properties": {
                "challenge": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "signed_challenge": {
                    "type": "string"
                }
            }
        },
        "models.Track": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "delete_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.TrackStatus"
                },
                "user_id": {
                    "type": "integer"
                },
                "visibility": {
                    "$ref": "#/definitions/models.TrackVisibility"
                }
            }
        },
        "models.TrackStatus": {
            "type": "string",
            "enum": [
                "enabled",
                "disabled"
            ],
            "x-enum-varnames": [
                "TrackStatusEnabled",
                "TrackStatusDisabled"
            ]
        },
        "models.TrackVisibility": {
            "type": "string",
            "enum": [
                "private",
                "public_r",
                "public_rw"
            ],
            "x-enum-varnames": [
                "TrackVisibilityPrivate",
                "TrackVisibilityPublicRead",
                "TrackVisibilityPublicWrite"
            ]
        },
        "models.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "public_key": {
                    "type": "string"
                },
                "salt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "key",
            "in": "query"
        },
        "TokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}